pyproject.toml
[project]
name = "poetry-lm"
version = "0.1.0"
description = "Russian poetry continuation model trained from scratch"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.2",
  "gradio>=5.23.0",
  "httpx>=0.28.1",
  "jsonlines>=4.0.0",
  "numpy>=2.1.0",
  "requests>=2.32.0",
  "ruaccent-predictor>=1.2.0",
  "sentencepiece>=0.2.0",
  "torch>=2.6.0",
  "tqdm>=4.67.0",
  "uvicorn>=0.30.0",
  "zstandard>=0.23.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.11.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
